refactor: 티커 도메인에서 티커 조회를 blockTicker도메인으로 이동
