+++ src/main/java/com/joojoo/api/ticker/presentation/controller/TickerController.java
@@ -1,10 +1,6 @@
 package com.joojoo.api.ticker.presentation.controller;
 
-import com.joojoo.global.common.response.detail.DailyBlockDetailsResponse;
-import com.joojoo.global.common.response.detail.count.TotalCountResponse;
 import com.joojoo.api.ticker.application.TickerService;
-import com.joojoo.api.ticker.application.port.in.GetTickerUseCase;
-import com.joojoo.api.ticker.presentation.dto.response.TickerSearchResponse;
 import com.joojoo.global.common.request.auth.CustomUserDetails;
 import com.joojoo.global.common.response.BaseResponse;
 import com.joojoo.global.common.response.ErrorResponse;
@@ -16,19 +12,18 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.web.bind.annotation.*;
-
-import java.time.LocalDate;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/ticker")
-@Tag(name = "Ticker API", description = "주식 티커 관련 API")
+@Tag(name = "Ticker API", description = "주식 마스터 티커 관련 API")
 public class TickerController {
     private final TickerService tickerService;
-    private final GetTickerUseCase getTickerUseCase;
 
     @Operation(summary = "주식 티커 추가 (검색 확인용)", description = "Redis에 새로운 주식 이름과 티커 심볼을 저장합니다.")
     @ApiResponses(value = {
@@ -50,53 +45,9 @@ public class TickerController {
         return BaseResponse.ok(name + " (" + ticker + ") 저장 성공!");
     }
 
-    @Operation(summary = "티커 검색", description = "쿼리를 통해 저장된 주식 티커를 검색합니다.")
-    @GetMapping("/search")
-    public BaseResponse<TickerSearchResponse> searchStock(
-            @Parameter(description = "검색어 (종목명)", required = true)
-            @RequestParam String query
-    ) {
-        return BaseResponse.ok(getTickerUseCase.search(query));
-    }
-
     @PostMapping("/load-Cache")
     public void dbToRedis(@AuthenticationPrincipal CustomUserDetails user){
         tickerService.loadTickersToCache(user.getUserId());
     }
 
-    @Operation(summary = "티커 상세 API", description = "특정 티커를 조회하여 Block, TradeLog에 작성한 티커를 가져옵니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "조회 성공",
-                    content = @Content(schema = @Schema(implementation = DailyBlockDetailsResponse.class)
-                    )
-            )
-    })
-    @GetMapping("/{tickerId}")
-    public BaseResponse<DailyBlockDetailsResponse> getTickerDetail(
-            @AuthenticationPrincipal CustomUserDetails user,
-            @PathVariable Long tickerId,
-            @RequestParam(required = false) @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate lastDate
-    ){
-        return BaseResponse.ok(getTickerUseCase.getTickerDetail(user.getUserId(), tickerId, lastDate));
-    }
-
-    @Operation(summary = "Ticker 상세 페이지 블럭 개수 API", description = "티커 상세 페이지 위에 티커 이름 하고 수량을 조회하는 API입니다.")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200", description = "조회 성공",
-                    content = @Content(schema = @Schema(implementation = TotalCountResponse.class)
-                    )
-            ),
-            @ApiResponse(
-                    responseCode = "404",
-                    description = "티커를 찾을 수 없습니다.",
-                    content = @Content(schema = @Schema(implementation = ErrorResponse.class))
-            )
-    })
-    @GetMapping("/{tickerId}/count")
-    public BaseResponse<TotalCountResponse> getTickerDetailCount(
-            @AuthenticationPrincipal CustomUserDetails user,
-            @PathVariable Long tickerId
-    ){
-        return BaseResponse.ok(getTickerUseCase.getTickerDetailCount(user.getUserId(), tickerId));
-    }
 }
